requirements.txt
Django==3.1.7
djangorestframework>=3.10.0
django-extensions
django-cors-headers
Pillow==8.1.2
requests==2.25.1
Unidecode==1.2.0
mysqlclient==2.0.3
